Add missing license header. Fix typo in class name.
