+++ oak-run/src/main/java/org/apache/jackrabbit/oak/index/IndexCommand.java
@@ -31,7 +31,7 @@ import org.apache.jackrabbit.oak.api.CommitFailedException;
 import org.apache.jackrabbit.oak.index.async.AsyncIndexerLucene;
 import org.apache.jackrabbit.oak.index.indexer.document.DocumentStoreIndexer;
 import org.apache.jackrabbit.oak.index.indexer.document.flatfile.FlatFileStore;
-import org.apache.jackrabbit.oak.plugins.index.FormatingUtils;
+import org.apache.jackrabbit.oak.plugins.index.FormattingUtils;
 import org.apache.jackrabbit.oak.plugins.index.MetricsFormatter;
 import org.apache.jackrabbit.oak.plugins.index.importer.IndexDefinitionUpdater;
 import org.apache.jackrabbit.oak.run.cli.CommonOptions;
@@ -264,7 +264,7 @@ public class IndexCommand implements Command {
         log.info("Indexing completed for indexes {} in {} ({} ms) and index files are copied to {}",
                 extendedIndexHelper.getIndexPaths(), reindexWatch, reindexWatch.elapsed(TimeUnit.MILLISECONDS), IndexCommand.getPath(destDir));
         log.info("[TASK:REINDEX:END] Metrics: {}",  MetricsFormatter.newBuilder()
-                .add("duration", FormatingUtils.formatToSeconds(reindexWatch))
+                .add("duration", FormattingUtils.formatToSeconds(reindexWatch))
                 .add("durationSeconds", reindexWatch.elapsed(TimeUnit.SECONDS))
                 .build());
         return destDir;
